+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Provide default encoding aliases (#416)
+- Add `get_encoding_by_alias` method to retrieve an encoding by alias
+
 ## [2.2.2] - 2024-02-17
 
 ### Changed